Dev: unittest: move tests to test_healthcheck
