+++ data-manifest
@@ -186,6 +186,7 @@ test/unittests/test_crashtest_task.py
 test/unittests/test_crashtest_utils.py
 test/unittests/test_gv.py
 test/unittests/test_handles.py
+test/unittests/test_healthcheck.py
 test/unittests/test_lock.py
 test/unittests/test_objset.py
 test/unittests/test_ocfs2.py
